+++ deno_dist/helper/ssg/index.ts
@@ -1,11 +1,10 @@
 import { replaceUrlParam } from '../../client/utils.ts'
 import type { Context } from '../../context.ts'
-import { filterStaticGenerateRoutes } from '../../helper/dev/index.ts'
 import type { Hono } from '../../hono.ts'
 import type { Env, MiddlewareHandler, Schema } from '../../types.ts'
 import { bufferToString } from '../../utils/buffer.ts'
 import { getExtension } from '../../utils/mime.ts'
-import { joinPaths, dirname } from './utils.ts'
+import { joinPaths, dirname, filterStaticGenerateRoutes } from './utils.ts'
 
 export const X_HONO_SSG_HEADER_KEY = 'x-hono-ssg'
 export const X_HONO_DISABLE_SSG_HEADER_KEY = 'x-hono-disable-ssg'
